Add warning for when using 'Compression.Pack200'
The compression stream that the 'Pack200' class uses has been removed since Java 14, so it seems apt to make sure to warn any potential users that usage of the it may fail at runtime, at compile time using an OptIn annotation. This annotation may be replaced with a Deprecated annotation with a level of ERROR, as the current way requires the user to at least the documentation to understand *why* it might be bad to use the class, but we'll see.

Signed-off-by: Oliver Berg <oliver@berg.moe>